requirements.txt
pdfplumber>=0.11.4
PyMuPDF>=1.25.0
pandas>=2.2.3
openpyxl>=3.1.5
reportlab>=4.2.5
python-dateutil>=2.9.0